Move Activable and Deletable intefaces
