+++ src/Payutc/OnyxBundle/Entity/TicketRepository.php
@@ -5,7 +5,7 @@ namespace Payutc\OnyxBundle\Entity;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\NoResultException;
 
-use Payutc\OnyxBundle\Entity\Deletable\DeletableEntityRepositoryInterface;
+use Payutc\OnyxBundle\Entity\Base\Deletable\DeletableEntityRepositoryInterface;
 
 /**
  * TicketRepository
